version 4.0.0

+++ server/server.js
@@ -2,37 +2,50 @@
 require('dotenv').config()
 
 // import stuff
-const express    = require('express')
-const app        = express()
-const path       = require('path')
-const bodyparser = require('body-parser')
-const logit      = require('./logit')
-
-// connect to database
-const db = require('mowr')(process.env.MOVIEDB_URI)
+const express = require('express')
+const app     = express()
+const path    = require('path')
+const charset = require('./charset')
+const cors    = require('./cors')
+const auth    = require('./auth')
+const logit   = require('./logit')
+const error   = require('http-errors')
+const jwt     = require('jwt-simple')
 
 // get movie collection
-const movies = db.get('movies')
+const movies = require('mowr')(process.env.MOVIEDB).get('movies')
 
-// create application/json parser
-app.use(bodyparser.json())
+// use json parser
+app.use(express.json())
 
-// enable CORS
-app.use((req, res, next) => {
-  res.header('Access-Control-Allow-Origin', '*')
-  res.header('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept')
-  res.header('Access-Control-Allow-Methods', 'GET, PUT, POST, DELETE, OPTIONS')
-  next()
-})
-
-// set path
+// set paths
 app.use(express.static(path.join(__dirname, '/')))
 app.use(express.static(path.join(__dirname, '/../')))
 app.use(express.static(path.join(__dirname, '/../dist/')))
 
+app.use(charset)
+app.use(cors)
+app.use(auth)
+
+// set secret string
+app.set('jwtsecret', process.env.JWTSECRET)
+
 // routes
-app.get('/', (req, res) => {
-  res.sendFile('index.html')
+app.get('/', (req, res) => res.sendFile('index.html'))
+
+app.post('/login', (req, res, next) => {
+  const role = req.body.pass === process.env.READ ? 'read' : req.body.pass === process.env.EDIT ? 'edit' : undefined
+
+  if (role) {
+    // token expires after 1 day
+    const token = jwt.encode({
+      exp: Date.now() + (1 * 24 * 60 * 60 * 1000),
+      role
+    }, app.get('jwtsecret'))
+    logit(`Someone logged in as ${role}`)
+    res.json({token})
+  }
+  else next(error(401, `Inloggen mislukt`))
 })
 
 app.get('/all', (req, res, next) => {
@@ -42,31 +55,51 @@ app.get('/all', (req, res, next) => {
 })
 
 app.post('/', (req, res, next) => {
-  movies.insertOne(req.body)
-  .then(x => res.json(x), logit(`POST --> ${req.body.title}`))
-  .catch(next)
+  if (req.role === 'edit') {
+    if (req.body && req.body.title) {
+      movies.insertOne(req.body)
+      .then(x => res.json(x), logit(`POST --> ${req.body.title}`))
+      .catch(next)
+    }
+    else logit('error post...req.headers:', req.headers)
+  }
+  else next(error(403, 'You cannot pass! I am a servant of the Secret Fire, wielder of the flame of Anor. The dark fire will not avail you, flame of Udûn. Go back to the Shadow! YOU! SHALL NOT! PASS!'))
 })
 
 app.put('/:id', (req, res, next) => {
-  movies.updateOne(req.params.id, {$set: req.body})
-  .then(x => res.json(x), logit(`PUT --> ${req.body.title}`))
-  .catch(next)
+  if (req.role === 'edit' || req.query.update) {
+    if (req.body && req.params.id) {
+      movies.updateOne(req.params.id, {$set: req.body})
+      .then(x => res.json(x), logit(`PUT --> ${req.body.title}`))
+      .catch(next)
+    }
+    else logit('error put...req.headers:', req.headers)
+  }
+  else next(error(403, 'You cannot pass! I am a servant of the Secret Fire, wielder of the flame of Anor. The dark fire will not avail you, flame of Udûn. Go back to the Shadow! YOU! SHALL NOT! PASS!'))
 })
 
 app.delete('/:id/:title', (req, res, next) => {
-  movies.deleteOne(req.params.id)
-  .then(() => res.json(true), logit(`DELETE --> ${req.params.title}`))
-  .catch(next)
+  if (req.role === 'edit') {
+    if (req.params.id && req.params.title) {
+      movies.deleteOne(req.params.id)
+      .then(() => res.json(true), logit(`DELETE --> ${req.params.title}`))
+      .catch(next)
+    }
+    else logit('error delete...req.headers:', req.headers)
+  }
+  else next(error(403, 'You cannot pass! I am a servant of the Secret Fire, wielder of the flame of Anor. The dark fire will not avail you, flame of Udûn. Go back to the Shadow! YOU! SHALL NOT! PASS!'))
 })
 
 // error handler
 app.use((err, req, res, next) => {
-  logit(err)
+  logit(`(${err.statusCode}) ${err}`)
+  if (!(err.statusCode === 401 || err.statusCode === 403)) logit(err)
+  
   if (!err.statusCode) err.statusCode = 500
   res.status(err.statusCode).json(err)
 })
 
 // start server
-app.listen(8088, () => {
-  logit(`Oscar's small Movie app is listening at port 8088`)
+app.listen(11188, () => {
+  logit(`Oscar's small Movie app is listening at port 11188`)
 })
